Split `<Report />`
