+++ app/report/types.ts
@@ -1,0 +1,1 @@
+export type CountryCode = "us" | "fr";
